fix: remove func to download
