+++ now/data_loading/data_loading.py
@@ -1,26 +1,17 @@
-import base64
 import os
-import pathlib
-import pickle
 import uuid
 from collections import defaultdict
 from copy import deepcopy
-from os.path import join as osp
 
 from docarray import Document, DocumentArray
 
 from now.app.base.app import JinaNOWApp
-from now.constants import (
-    BASE_STORAGE_URL,
-    DEMO_DATASET_DOCARRAY_VERSION,
-    DatasetTypes,
-    Modalities,
-)
+from now.constants import DatasetTypes
 from now.data_loading.elasticsearch import ElasticsearchExtractor
 from now.demo_data import DemoDatasetNames
 from now.log import yaspin_extended
 from now.now_dataclasses import UserInput
-from now.utils import download, sigmap
+from now.utils import sigmap
 
 
 def load_data(app: JinaNOWApp, user_input: UserInput) -> DocumentArray:
@@ -189,53 +180,6 @@ def _list_files_from_s3_bucket(app: JinaNOWApp, user_input: UserInput) -> Docume
                         break
 
     return DocumentArray(docs)
-
-
-def fetch_da_from_url(
-    url: str, downloaded_path: str = '~/.cache/jina-now'
-) -> DocumentArray:
-    data_dir = os.path.expanduser(downloaded_path)
-    if not os.path.exists(osp(data_dir, 'data/tmp')):
-        os.makedirs(osp(data_dir, 'data/tmp'))
-    data_path = (
-        data_dir
-        + f"/data/tmp/{base64.b64encode(bytes(url, 'utf-8')).decode('utf-8')}.bin"
-    )
-    if not os.path.exists(data_path):
-        download(url, data_path)
-
-    try:
-        da = DocumentArray.load_binary(data_path)
-    except pickle.UnpicklingError:
-        path = pathlib.Path(data_path).expanduser().resolve()
-        os.remove(path)
-        download(url, data_path)
-        da = DocumentArray.load_binary(data_path)
-    return da
-
-
-def get_dataset_url(dataset: str, output_modality: Modalities) -> str:
-    data_folder = None
-    docarray_version = DEMO_DATASET_DOCARRAY_VERSION
-    if output_modality == Modalities.IMAGE:
-        data_folder = 'jpeg'
-    elif output_modality == Modalities.TEXT:
-        data_folder = 'text'
-    elif output_modality == Modalities.MUSIC:
-        data_folder = 'music'
-    elif output_modality == Modalities.VIDEO:
-        data_folder = 'video'
-    elif output_modality == Modalities.TEXT_AND_IMAGE:
-        data_folder = 'text-image'
-    if output_modality not in [
-        Modalities.MUSIC,
-        Modalities.VIDEO,
-        Modalities.TEXT_AND_IMAGE,
-    ]:
-        model_name = 'ViT-B32'
-        return f'{BASE_STORAGE_URL}/{data_folder}/{dataset}.{model_name}-{docarray_version}.bin'
-    else:
-        return f'{BASE_STORAGE_URL}/{data_folder}/{dataset}-{docarray_version}.bin'
 
 
 def get_s3_bucket_and_folder_prefix(user_input: UserInput):
