+++ now/data_loading/migrate_datasets.py
@@ -1,14 +1,19 @@
+import base64
 import multiprocessing as mp
+import os
+import pathlib
+import pickle
 import sys
+from os.path import join as osp
 from typing import List
 
 from docarray import Document, DocumentArray, dataclass
 from docarray.typing import Blob, Image, Text, Video
 from tqdm import tqdm
 
-from now.constants import Modalities
-from now.data_loading.utils import fetch_da_from_url, get_dataset_url
+from now.constants import BASE_STORAGE_URL, DEMO_DATASET_DOCARRAY_VERSION, Modalities
 from now.demo_data import AVAILABLE_DATASET, DemoDataset
+from now.utils import download
 
 current_module = sys.modules[__name__]
 
@@ -167,6 +172,53 @@ def convert_dataset(dataset: DemoDataset, modality: Modalities, num_workers: int
         show_progress=True,
     )
     print(f'  Saved {dataset.name} docs to hubble ...')
+
+
+def fetch_da_from_url(
+    url: str, downloaded_path: str = '~/.cache/jina-now'
+) -> DocumentArray:
+    data_dir = os.path.expanduser(downloaded_path)
+    if not os.path.exists(osp(data_dir, 'data/tmp')):
+        os.makedirs(osp(data_dir, 'data/tmp'))
+    data_path = (
+        data_dir
+        + f"/data/tmp/{base64.b64encode(bytes(url, 'utf-8')).decode('utf-8')}.bin"
+    )
+    if not os.path.exists(data_path):
+        download(url, data_path)
+
+    try:
+        da = DocumentArray.load_binary(data_path)
+    except pickle.UnpicklingError:
+        path = pathlib.Path(data_path).expanduser().resolve()
+        os.remove(path)
+        download(url, data_path)
+        da = DocumentArray.load_binary(data_path)
+    return da
+
+
+def get_dataset_url(dataset: str, output_modality: Modalities) -> str:
+    data_folder = None
+    docarray_version = DEMO_DATASET_DOCARRAY_VERSION
+    if output_modality == Modalities.IMAGE:
+        data_folder = 'jpeg'
+    elif output_modality == Modalities.TEXT:
+        data_folder = 'text'
+    elif output_modality == Modalities.MUSIC:
+        data_folder = 'music'
+    elif output_modality == Modalities.VIDEO:
+        data_folder = 'video'
+    elif output_modality == Modalities.TEXT_AND_IMAGE:
+        data_folder = 'text-image'
+    if output_modality not in [
+        Modalities.MUSIC,
+        Modalities.VIDEO,
+        Modalities.TEXT_AND_IMAGE,
+    ]:
+        model_name = 'ViT-B32'
+        return f'{BASE_STORAGE_URL}/{data_folder}/{dataset}.{model_name}-{docarray_version}.bin'
+    else:
+        return f'{BASE_STORAGE_URL}/{data_folder}/{dataset}-{docarray_version}.bin'
 
 
 def main():
